refactor(mdb): Move error to separate module
To make it easier to separate the refactoring from the changes in #116.
Additionally this should may help by creating a smaller module with
clear safe and unsafe interfaces; i.e. to audit the safety of the error
bindings we need only look at the much smaller error module.
