+++ apps-aarch64.checksum
@@ -1,7 +1,7 @@
 9d66c3ff02cf209762e13cd5d7cd5d91b6da23c2  target-aarch64/acap/Challenge_Build_Tools_1_0_0_all.eap
 12a30609c7fa463e1e8d0405665d2aa17d4349c4  target-aarch64/acap/axstorage_example_0_0_0_aarch64.eap
 6db963922ca2ceb25715bbdb6445b504c307d432  target-aarch64/acap/bounding_box_example_0_0_0_aarch64.eap
-115a5c3422f536d046e3a17774969c95ae1873c2  target-aarch64/acap/consume_analytics_metadata_0_0_0_aarch64.eap
+82942aa6caf385ec742a9f126078cd3a4c20a76d  target-aarch64/acap/consume_analytics_metadata_0_0_0_aarch64.eap
 ad1142937ed6e2d4f2a1f4cb0fdae721a0744750  target-aarch64/acap/embedded_web_page_0_0_0_aarch64.eap
 f47872232e0f2cd4ace9e2e77b07cafb431cbe73  target-aarch64/acap/hello_world_0_0_0_aarch64.eap
 75818664787ff2ff7dbdd32fadfe20c23974c867  target-aarch64/acap/inspect_env_0_0_0_aarch64.eap
